Created the Game, Event and Graphic Service and Service Locator
Created the Graphic Service header and source files and declared and defined necessary functions to display game window.

Created the Event Service header and source files and declared and defined the necessary functions to process user input like mouse clicks and keyboard presses and to execute appropriate actions based on the input.

Created the Service Locator header and source files and declared and defined necessary Service pointers and functions to call the relevant functions of the different Service classes.

Linked the Graphic and Event Service to the Service Locator by including the Services' header files in the ServiceLocator header file.

Created the Game Service header and source files to call the relevant functions of the Service Locator which in turn calls the functions of the Service classes.

Called the appropriate functions of the Game Service from the main.cpp file.