Implemented Spawning, Movement and Destruction of Ducks
Created the Duck Service, Controller, Model and View header and source files.

Declared and defined necessary functions and variables in DuckService header and source files to spawn and destroy ducks.

Declared and defined necessary functions and pointers in the DuckController header and source files to call the DuckModel and DuckView functions and to implement movement of ducks and destruction of ducks when mouse clicking on the ducks.

Declared and defined necessary functions and variables in DuckModel header and source files to set and retrieve the duck position and movement direction.

Declared and defined necessary functions and variables in the DuckView header and source files to display the Duck image in the game window.