Created the Player Service, Controller and Model
Declared and defined the necessary functions and PlayerController pointer in the PlayerService header and source files to call the appropriate PlayerController functions.

Declared and defined the necessary functions and Player Model pointer in the PlayerController header and source files to increase Player score, decrease number of bullets left and to call appropriate PlayerModel functions.

Declared and defined the necessary functions and variables in the PlayerModel header and source files to set the number of bullets and the player score.

Linked the Player Service to the Service Locator.