Created the UI Service, Gameplay UI, Wave Screens and Implemented difficulty increase per wave
Declared and defined necessary variables and functions in the TextView header and source files to set properties of text such as font type, size, position and print it in game window.

Declared and defined necessary functions in the GameplayUI header and source files to show the player score and number of bullets left in the game window by calling the appropriate TextView functions.

Declared and defined the necessary functions in the WaveUIController header and source files to display which wave will be starting in the game window by calling the appropriate TextView functions.

Declared and defined necessary functions in the UIService to call the appropriate GameplayUIController and WaveUIController functions.

Linked UIService to the Service Locator.

Implemented increasing duck speeds with every new wave in the DuckService and a maximum of 3 waves in the WaveUIController.