Implemented Round replay on Round Loss and Timer for Round
Declared and defined appropriate functions in the WaveUIController header and source files to replay the same wave/round when the player has lost the round.

Declared and defined necessary functions and variables in the GameplayUIController header and source files to implement timer for each round.

Made changes to the DuckService header and source files to implement specified number of ducks to be spawned each round and increasing that number with each subsequent round.