Implemented Shot Ducks Falling
Made changes to the DuckView and DuckController header and source files by declaring and defining necessary functions to set and check the state of Duck (FLYING or DEAD) and to call the appropriate functions accordingly.

Made change to the GameplayUIController by making sure Radial button can change bullet type to Radial only if Radial bullet has not been used.